requirements.txt
sentence-transformers
scikit-learn
PyPDF2
python-docx
